requirements.txt
apache-airflow==2.7.3
pandas==2.0.3
praw==7.7.1
